Develop

+++ src/IdentityServer.MultiTenant/IdentityServer.MultiTenant/Models/TenantInfoModel.cs
@@ -7,10 +7,10 @@ namespace IdentityServer.MultiTenant.Models
 {
     public class TenantInfoModel
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string GuidId { get; set; }
         public string Identifier { get; set; }
-        public int TenantDomainId { get; set; }
+        public Int64 TenantDomainId { get; set; }
         /// <summary>
         /// 启用状态
         /// 1：启用，0：禁用
